moved legacy crawlers not based on VoxxrinCrawler to legacy_crawler.js
